+++ src/components/cafe/Order/ListItem/OrderListItemView.ts
@@ -1,6 +1,6 @@
 import { Order } from '@/domain';
 
-export const template = (name: string, order: Order) => String.raw`
+export const orderListItemView = (name: string, order: Order) => String.raw`
 <div class='table-row order-list-item' data-order-id='${order.getId()}'>
     <div class='cell' data-title='No'></div>
     <div class='cell' data-title='메뉴명'>${name}</div>
