+++ src/components/cafe/Order/OrderListView.ts
@@ -1,4 +1,4 @@
-export const template = String.raw`
+export const orderListView = String.raw`
 <div class='order-list'>
   <h1>주문 목록</h1>
   <div class='order-button-area'>
