+++ src/components/cafe/Serving/ServedView.ts
@@ -1,4 +1,4 @@
-export const template = () => String.raw`
+export const servedView = () => String.raw`
 <div class='cafe-serving'>
   <span class="maked-list-view">
     <button>현재까지 서빙된 커피 확인하기</button>
