+++ src/components/index.ts
@@ -5,6 +5,6 @@ export * from '@/components/cafe/Menu/Menu';
 export * from '@/components/cafe/Modal/Modal';
 
 export * from '@/components/cafe/Order/OrderList';
-export * from '@/components/cafe/Order/OrderListItem';
+export * from '@/components/cafe/Order/ListItem/OrderListItem';
 
 export * from '@/components/cafe/Serving/Served';
